+++ tools/src/com.oracle.truffle.tools.profiler/src/com/oracle/truffle/tools/profiler/CPUTracer.java
@@ -47,7 +47,6 @@ import com.oracle.truffle.api.instrumentation.StandardTags.RootTag;
 import com.oracle.truffle.api.instrumentation.StandardTags.StatementTag;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument.Env;
-import com.oracle.truffle.api.nodes.NodeCost;
 import com.oracle.truffle.api.source.SourceSection;
 import com.oracle.truffle.tools.profiler.impl.CPUTracerInstrument;
 import com.oracle.truffle.tools.profiler.impl.ProfilerToolFactory;
@@ -60,7 +59,8 @@ import com.oracle.truffle.tools.profiler.impl.ProfilerToolFactory;
  * The tracer counts how many times each of the elements of interest (e.g. functions, statements,
  * etc.) are executed.
  * <p>
- * Usage example: {@snippet file = "com/oracle/truffle/tools/profiler/CPUTracer.java" region =
+ * Usage example:
+ * {@snippet file = "com/oracle/truffle/tools/profiler/CPUTracer.java" region =
  * "CPUTracerSnippets#example"}
  *
  * @since 0.30
@@ -302,10 +302,6 @@ public final class CPUTracer implements Closeable {
             }
         }
 
-        @Override
-        public NodeCost getCost() {
-            return NodeCost.NONE;
-        }
     }
 
     static ProfilerToolFactory<CPUTracer> createFactory() {